Move chart bundler in web worker
