+++ package.json
@@ -17,7 +17,7 @@
     "bootstrap": "^4.5.0",
     "classnames": "^2.2.6",
     "clipboard": "^2.0.8",
-    "comlink": "^4.3.0",
+    "comlink": "^4.3.1",
     "d3": "^7.2.0",
     "dayjs": "^1.8.33",
     "fast-deep-equal": "^3.1.3",
@@ -41,7 +41,7 @@
     "sparql-http-client": "^2.4.0",
     "sparqljs": "^3.5.1",
     "typeface-inter": "^3.12.0",
-    "worker-loader": "^3.0.2"
+    "worker-loader": "=2.0.0"
   },
   "scripts": {
     "format": "prettier --write src",
